Extract shared Row/Summary/Band types into _rows.py, fixing circular dependency
snapshot.py and _health_ranking.py shared a circular dependency: snapshot
defined Band/Row/Summary that _health_ranking imported, while snapshot
imported band_from_health/compute_team_row/rank_rows from _health_ranking.
The cycle was papered over with a deferred import (noqa: E402).

Fix: lift Band, Row, Summary, band_color into a leaf module (_rows.py)
with zero dependencies on the rest of the metrics package. Both snapshot
and _health_ranking import from _rows. No more deferred imports.
